refactor(samples): the whole binary lives in the library
A process shell that parses, runs and prints is a piece of the program
no test can drive without spawning one - and the parts most worth
testing, the refusals, are exactly the parts a spawned process makes
hardest to inspect. Coverage said so first: twenty-four lines of main
with no test able to reach them.

Main is now a single expression, matching the sample beside it, and the
command line is driven directly by a test that asserts each exit code.
